tidymodels: Easily Install and Load the 'Tidymodels' Packages

The tidy modeling "verse" is a collection of packages for modeling and statistical analysis that share the underlying design philosophy, grammar, and data structures of the tidyverse.

See also

Author

Maintainer: Max Kuhn max@posit.co (ORCID)

Authors:

Hadley Wickham hadley@posit.co

Other contributors:

Posit Software, PBC (03wc8by49) [copyright holder, funder]