Facilities for loading and updating other packages

`tidymodels` contains a set of tags for specific topics in modeling and analysis. These can be used to load packages in groups.

tag_show()

tag_attach(tag)

tag_update(tag)

Arguments

tag
A character strong for the tag. Current options are 'Bayesian analysis', 'deep learning', 'feature engineering', 'parallel processing', 'resampling', 'text analysis', and 'time series'.

Examples

tag_show()

#> All tags:
#> 
#> Bayesian analysis: 'tidyposterior', 'tidybayes'
#> deep learning: 'keras'
#> feature engineering: 'keras', 'textrecipes'
#> parallel processing: 'furrr'
#> resampling: 'infer', 'tidyposterior'
#> text analysis: 'tidytext', 'keras', 'textrecipes'
#> time series: 'timetk', 'tidyquant', 'sweep'
#> extra recipes: 'embed', 'textrecipes', 'themis'
#> extra models: 'discrim'